+++ Packages/webxr/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Using `frame.fillPoses` and `frame.fillJointRadii` instead of `frame.getJointPose` for XRHand.
+- Camera matrices handling.
+
+### Removed
+- Matrix utilities files.
 
 ## [0.5.2] - 2020-01-16
 ### Added